style: change component name and filename
